Update gr.yml
Created an account on github to contribute to this translation.
Based on the en.yml, I made the following changes to this document 
*Improved the vocabulary. Used words that matched the english definition better than the previous ones or made no sense.
*Fixed spelling mistakes.
*Fixed grammatical errors.
*Reverted phrases such as "Ground control to Major Tom", back to english, since they made no sense if they were translated to Greek.

Please note that the translation is for the most part complete, but there need to be a few more changes in order for it to be ready to be used on the website. Please contact me if you need more clarifications on the changes I have made.