First translation attempt
Translated everything except the payment procedures.
Small inconsistencies about how to translate Support FetLife. Translated as naturally as possible except when is refered as a product, then is translated as "Apoyo a Fetlife" with double quotation marks.
All the lines translated you as 'Usted'(formal you) even if sometimes sounds a bit too forced. Spanish is no that polite.
Several pop culture references are removed because they are not well known in the spanish world.
The joke about using canola oil as lube have a little bit of negative connotations but only in Spain. In the eighties a big shipment of adulterated canola oil was sold by unexcrupulous businessmen and a big number of people died or got sequela for life.