[ES] Remove unnecessary abbreviations (Ud.)
In Spanish it is recommended not to use abbreviations save in a small
set of circumstances (generally lists, forms of addressing, or where
space is at a premium). The only cases in need of change were instances
of Ud. (usted), bringing them in line with the majority of ocurrences of
that word.