Fix grammatical mistakes and style (DE)
Explanations for some of the corrections:

The dash is wrong between two words when the first word has a
'Fugen-S' which implies it to be genitive. Thus I removed it.

'URL' as in 'Website-URL' can have either female or male grammatical
gender as per the Duden. Male, however, is far more uncommon, thus
the change of the article from 'Der' to 'Die'.

The apostroph is not necessary in the imperative when there's no
letter missing.

'zum Bankkonto gehörig' is a participle construction which makes it
generally a bit hard to read and understand. The genitive will already
be enough to show posessiveness.