Change capitalisation of informal honorifcs
'du' was written with a capital D until the spelling reform of 1996.
Since then it is usually written with a small d. There are exceptions
where you *can* write it still with a capital D, if that's your thing,
but those only apply to 1-to-1 communication like in SMS, E-Mail, etc.